Separate out logic for Crown daily list and Magistrates public list
