+++ src/main/java/uk/gov/hmcts/reform/pip/channel/management/services/filegeneration/MagistratesPublicListFileConverter.java
@@ -4,11 +4,9 @@ import com.fasterxml.jackson.databind.JsonNode;
 import org.springframework.stereotype.Service;
 import org.thymeleaf.spring5.SpringTemplateEngine;
 import uk.gov.hmcts.reform.pip.channel.management.config.ThymeleafConfiguration;
-import uk.gov.hmcts.reform.pip.model.publication.ListType;
+import uk.gov.hmcts.reform.pip.channel.management.services.helpers.listmanipulation.MagistratesPublicListHelper;
 
 import java.util.Map;
-
-import static uk.gov.hmcts.reform.pip.channel.management.services.helpers.listmanipulation.CrimeListHelper.preprocessArtefactForCrimeListsThymeLeafConverter;
 
 @Service
 public class MagistratesPublicListFileConverter implements FileConverter {
@@ -16,7 +14,7 @@ public class MagistratesPublicListFileConverter implements FileConverter {
     public String convert(JsonNode artefact, Map<String, String> artefactValues, Map<String, Object> language) {
         SpringTemplateEngine templateEngine = new ThymeleafConfiguration().templateEngine();
         return templateEngine.process("magistratesPublicList.html",
-                                      preprocessArtefactForCrimeListsThymeLeafConverter(
-                                          artefact, artefactValues, language, ListType.MAGISTRATES_PUBLIC_LIST));
+                                      MagistratesPublicListHelper.preprocessArtefactForThymeLeafConverter(
+                                          artefact, artefactValues, language));
     }
 }
